Implement safe_write
Despite the documentation hinting Unix.write would in all cases write all data,
or error out, the implementation shows on EAGAIN or EWOULDBLOCK the returned
number of written bytes might be smaller than the intended number of bytes to be
written.

This commit introduces a 'safe_write' function which uses Unix.single_write,
calling it multiple times until all data is written, or an error occurs.