Add min_blocks argument to compact procedure
This allows to limit the number of hole punch calls on compaction, by only
allowing 'large' holes to be punched.

When using the default value, 0, a hole will always be punched, even though it's
smaller than a single block. This is different compared to the previous
behaviour (which required at least a single block to be punched).

Some test cases are added as well.