Don't set POSIX_FADV_RANDOM on read FD
Setting the POSIX_FADV_RANDOM advice on a file descriptor seems to have an
enormous (negative) performance impact in our access pattern. I didn't look into
the Linux kernel sources too closely, but it looks like setting this advice
causes the page cache to be populated only with the requested data. Since we
always need 2 read calls (one to figure out the entry size, once to read the
actual entry data), this causes 2 reads to the block layer, whilst without the
advice the entry data will be prefetched (it's contained in the same page most
of the time anyway).