Cache recently-written slab entries
This commit adds a 32-entry cache to Flog, which keeps track of the most
recently written entries. This has quite some impact on the number of syscalls
required, as well as the number of deserializations: when running the benchmark
with bench-size 1000, the Flog0 implementation uses 37468 read(2) calls and
20738 corresponding lseek(2) calls, so before this commit Flog would use 37468
pread(2) calls. Now only 11413 pread(2) calls are executed thanks to some basic
caching.