benchmark: Add --min-blocks flag
Using the --min-blocks flag one can set the minimal number of
consecutive blocks to punch when using the --punch command. This value
is passed as the min_blocks optional argument of Flog.compact.

The default value of this flag is set to 1, which should speed up
compaction considerably compared to the API default of 0, which writes
zeros to disk when punching a hole smaller than the FS block size (and
as such doesn't save any disk space).