Add optional progress callback to compact
This commit adds an optional "progress_cb" argument to Log.compact,
by which a callback can be provided which will be called with both the
total database file size as well as the current index being handled
during compaction. This way, compaction progress can be displayed to
the end-user.

Note (in the Flog implementation) this progress does not denote total
time to completion, it merely provides a way to measure which part of
the database file has been handled, and how much is left to do. This
does not give any guarantees regarding total time to completion.