benchmark: Add compaction progress display
This commit uses the progress callback introduced in c2cab19435862 to
provide some feedback on 'hole punching compaction' progress.

Progress is displayed in steps of 5%. As noted in the commit message of
c2cab19435862, this does not necessarily indicate the amount of time
until completion of the operation, only which part of the database file
has been processed.

See: c2cab19435862c25052fd6d522b13ebef7cda8e7