Add UUID generation code
This commit adds some simple bindings to libuuid to generate UUIDs,
which will be used to 'tag' database files in a multi-spindle setup in
an attempt to trick the universe.

This commit includes an ugly build fix for recent versions of
ld/binutils, until ocamlbuild has been adapted to cope with these
changes. See comments inline.

Change-Id: I75896d347969fd31efcc4be2d96fb8843982a0ce