Rework into monadic style
This commit changes Baardskeerder all over the place, changing the
Log interface to allow for various IO mechanisms, so we can create an
Lwt-backed Log implementation.

More information on the rationale behind this can be found in some blog
post on the Incubaid Research blog (once published...).

This commit adds a monadic interface to Log ('a m/bind/return/run),
converts the Tree module to use this new interface, and updates test
cases to compile and pass.

One drawback: the 'Rewrite' module can't be converted as-is, and as
such rewrite support is no longer provided, for now.

Change-Id: I47fc8ab5ef23f01b87cd8cdea83a243f6c79f50f