Add Mlog_lwt
This module is a simple wrapper around Mlog, but acting as if it's
backed by Lwt. It simply passes all calls to the pure Mlog
implementation it encapsulates, and lifts the result into an Lwt.t
value.

This is a simple test to check whether the changes introduced in
d98a85cf can actually work in an Lwt-based application.

Change-Id: Ib9e1233f726707d011b251f5ebde9135d61a97eb
See: d98a85cf66d06be3bb9d1c280706de9b10438769