Functorize Flog0
According to the new way of working, serialization implementations
should be functors over a Store.STORE module to result in a LOG
implementation.

This commit alters the Flog0 module to provide such a functor, and
updates the places where the plain Flog0 module was used before into
Flog0 applied to the Store.Sync module.

Change-Id: I3c057629f4b6d452517e331a41c9adac985cf7d3