Store: Don't wrap values in Memory and Sync store
Wrapping the values requires the usage of the `run` function, which is
rather ugly in case the user of the library uses a synchronous/blocking
backend (see `sync_demo.ml` in a following commit).

This commit alters the definitions in the Store module to expose the
internal monadic type of the Memory and Sync modules, which is the inner
type (S.('a m) = 'a).

`bind` becomes function application, `return` is the identity function
and so is `run`.

Change-Id: I94a0b2e50e0a179174da7db6a82f8f3c80e3835f